avalanche-api: rename to x/p/c
