+++ avalanche-api/src/lib.rs
@@ -1,7 +1,6 @@
-pub mod avm;
-pub mod eth;
+pub mod c;
 pub mod health;
 pub mod info;
 pub mod metrics;
 pub mod p;
-pub mod platform;
+pub mod x;
