+++ subnetctl/src/get_utxos/mod.rs
@@ -4,7 +4,7 @@ use clap::{Arg, Command};
 use log::info;
 use tokio::runtime::Runtime;
 
-use avalanche_api::platform;
+use avalanche_api::p;
 
 pub const NAME: &str = "get-utxos";
 
@@ -57,7 +57,7 @@ pub fn execute(opt: Option) -> io::Result<()> {
     let rt = Runtime::new().unwrap();
 
     let resp = rt
-        .block_on(platform::get_utxos(&opt.http_rpc_ep, &opt.paddr))
+        .block_on(p::get_utxos(&opt.http_rpc_ep, &opt.paddr))
         .expect("failed to get UTXOs");
     info!("get_utxos response: {:?}", resp);
